+++ app/src/main/java/uk/co/keepawayfromfire/screens/tasker/Consts.java
@@ -1,0 +1,10 @@
+package uk.co.keepawayfromfire.screens.tasker;
+
+/**
+ * Created by cj on 09/01/17.
+ */
+
+public class Consts {
+    public static final String ACTION_EDIT_SETTING
+            = "com.twofortyfouram.locale.intent.action.EDIT_SETTING";
+    public static final String EXTRA_BUNDLE = "com.twofortyfouram.locale.intent.extra.BUNDLE";}